Project2

+++ Project2/project2/Controllers/DonHangsController.cs
@@ -1,13 +1,9 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Threading.Tasks;
-using Microsoft.AspNetCore.Mvc;
+﻿using Microsoft.AspNetCore.Mvc;
 using Microsoft.AspNetCore.Mvc.Rendering;
 using Microsoft.EntityFrameworkCore;
 using Project2.Models;
 
-namespace Project2.Controllers
+namespace FastFood.Controllers
 {
     public class DonHangsController : Controller
     {
@@ -18,84 +14,69 @@ namespace Project2.Controllers
             _context = context;
         }
 
+        // 📌 Lấy role & customerId từ session
+        private string? GetRole() => HttpContext.Session.GetString("Role");
+        private int? GetCustomerId() => HttpContext.Session.GetInt32("CustomerId");
+
         // GET: DonHangs
         public async Task<IActionResult> Index()
         {
-            var qlbanDoAnContext = _context.DonHangs.Include(d => d.MaKhNavigation);
-            return View(await qlbanDoAnContext.ToListAsync());
+            var role = GetRole();
+            var customerId = GetCustomerId();
+
+            IQueryable<DonHang> query = _context.DonHangs
+                .Include(d => d.MaKhNavigation);
+
+            if (role == "Customer" && customerId != null)
+            {
+                query = query.Where(d => d.MaKh == customerId);
+            }
+
+            return View(await query.ToListAsync());
         }
 
         // GET: DonHangs/Details/5
         public async Task<IActionResult> Details(int? id)
         {
-            if (id == null)
-            {
-                return NotFound();
-            }
+            if (id == null) return RedirectToAction(nameof(Index));
 
             var donHang = await _context.DonHangs
                 .Include(d => d.MaKhNavigation)
                 .FirstOrDefaultAsync(m => m.MaDh == id);
-            if (donHang == null)
+
+            if (donHang == null) return RedirectToAction(nameof(Index));
+
+            // Khách không được xem đơn người khác
+            var role = GetRole();
+            var customerId = GetCustomerId();
+            if (role == "Customer" && donHang.MaKh != customerId)
             {
-                return NotFound();
-            }
-
-            return View(donHang);
-        }
-
-        // GET: DonHangs/Create
-        public IActionResult Create()
-        {
-            ViewData["MaKh"] = new SelectList(_context.KhachHangs, "MaKh", "MaKh");
-            return View();
-        }
-
-        // POST: DonHangs/Create
-        // To protect from overposting attacks, enable the specific properties you want to bind to.
-        // For more details, see http://go.microsoft.com/fwlink/?LinkId=317598.
-        [HttpPost]
-        [ValidateAntiForgeryToken]
-        public async Task<IActionResult> Create([Bind("MaDh,NgayDat,TrangThai,PhuongThucTt,TongTien,MaKh")] DonHang donHang)
-        {
-            if (ModelState.IsValid)
-            {
-                _context.Add(donHang);
-                await _context.SaveChangesAsync();
                 return RedirectToAction(nameof(Index));
             }
-            ViewData["MaKh"] = new SelectList(_context.KhachHangs, "MaKh", "MaKh", donHang.MaKh);
+
             return View(donHang);
         }
 
         // GET: DonHangs/Edit/5
         public async Task<IActionResult> Edit(int? id)
         {
-            if (id == null)
-            {
-                return NotFound();
-            }
+            if (GetRole() != "Admin") return Unauthorized();
+            if (id == null) return RedirectToAction(nameof(Index));
 
             var donHang = await _context.DonHangs.FindAsync(id);
-            if (donHang == null)
-            {
-                return NotFound();
-            }
-            ViewData["MaKh"] = new SelectList(_context.KhachHangs, "MaKh", "MaKh", donHang.MaKh);
+            if (donHang == null) return RedirectToAction(nameof(Index));
+
+            ViewData["MaKh"] = new SelectList(_context.KhachHangs, "MaKh", "HoTen", donHang.MaKh);
             return View(donHang);
         }
 
         // POST: DonHangs/Edit/5
-        // To protect from overposting attacks, enable the specific properties you want to bind to.
-        // For more details, see http://go.microsoft.com/fwlink/?LinkId=317598.
         [HttpPost]
         [ValidateAntiForgeryToken]
-        public async Task<IActionResult> Edit(int id, [Bind("MaDh,NgayDat,TrangThai,PhuongThucTt,TongTien,MaKh")] DonHang donHang)
+        public async Task<IActionResult> Edit(int id, DonHang donHang)
         {
-            if (id != donHang.MaDh)
-            {
-                return NotFound();
-            }
+            if (GetRole() != "Admin") return Unauthorized();
+            if (id != donHang.MaDh) return RedirectToAction(nameof(Index));
 
             if (ModelState.IsValid)
             {
@@ -106,36 +87,29 @@ namespace Project2.Controllers
                 }
                 catch (DbUpdateConcurrencyException)
                 {
-                    if (!DonHangExists(donHang.MaDh))
-                    {
-                        return NotFound();
-                    }
+                    if (!_context.DonHangs.Any(e => e.MaDh == donHang.MaDh))
+                        return RedirectToAction(nameof(Index));
                     else
-                    {
                         throw;
-                    }
                 }
                 return RedirectToAction(nameof(Index));
             }
-            ViewData["MaKh"] = new SelectList(_context.KhachHangs, "MaKh", "MaKh", donHang.MaKh);
+
+            ViewData["MaKh"] = new SelectList(_context.KhachHangs, "MaKh", "HoTen", donHang.MaKh);
             return View(donHang);
         }
 
         // GET: DonHangs/Delete/5
         public async Task<IActionResult> Delete(int? id)
         {
-            if (id == null)
-            {
-                return NotFound();
-            }
+            if (GetRole() != "Admin") return Unauthorized();
+            if (id == null) return RedirectToAction(nameof(Index));
 
             var donHang = await _context.DonHangs
                 .Include(d => d.MaKhNavigation)
                 .FirstOrDefaultAsync(m => m.MaDh == id);
-            if (donHang == null)
-            {
-                return NotFound();
-            }
+
+            if (donHang == null) return RedirectToAction(nameof(Index));
 
             return View(donHang);
         }
@@ -145,19 +119,22 @@ namespace Project2.Controllers
         [ValidateAntiForgeryToken]
         public async Task<IActionResult> DeleteConfirmed(int id)
         {
+            if (GetRole() != "Admin") return Unauthorized();
+
             var donHang = await _context.DonHangs.FindAsync(id);
             if (donHang != null)
             {
+                // Xóa chi tiết đơn hàng liên quan trước khi xóa đơn hàng
+                var chiTietDonHangs = _context.ChiTietDonHangs.Where(ct => ct.MaDh == id);
+                _context.ChiTietDonHangs.RemoveRange(chiTietDonHangs);
+
+                // Xóa đơn hàng
                 _context.DonHangs.Remove(donHang);
+
+                await _context.SaveChangesAsync();
             }
 
-            await _context.SaveChangesAsync();
             return RedirectToAction(nameof(Index));
-        }
-
-        private bool DonHangExists(int id)
-        {
-            return _context.DonHangs.Any(e => e.MaDh == id);
         }
     }
 }
